requirements.txt
streamlit
crewai
crewai_tools
litellm
google-generativeai>=0.8.4
python-dotenv
crewai[tools]>=0.100.0
pyaudio